fix issue #45, pre-process cert data - pre-process com.apple.security.pkcs1 PayloadContent data values - validate type: must be CFPropertyList::Blob or String - decode as required and parse the data using OpenSSL to validate
This will allow people to express their PayloadContent data in 3 flavours:

1. Base64 encoded CFPropertyList::Blob (preferred)

    - PayloadContent: !ruby/string:CFPropertyList::Blob |-
        LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUViRENDQTFTZ0F3SUJB
        Z0lRVFY4c05BaXlUQ0ROYlZCK0pFM0o3REFOQmdrcWhraUc5dzBCQVFVRkFE
        Q0IKcVRFTE1Ba0dBMVVFQmhNQ1ZWTXhGVEFUQmdOVkJBb1RESFJvWVhkMFpT
        ...

2. Base64 encoded Binary string

    - PayloadContent: !binary |-
        LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUVJRENDQXdpZ0F3SUJB
        Z0lRTkU3VlZ5RFY3ZXhKOUMvT045c3JiVEFOQmdrcWhraUc5dzBCQVFVRkFE
        Q0IKcVRFTE1Ba0dBMVVFQmhNQ1ZWTXhGVEFUQmdOVkJBb1RESFJvWVhkMFpT
        ...

3. PEM formatted String

    - PayloadContent: |
        -----BEGIN CERTIFICATE-----
        MIIEtzCCA5+gAwIBAgIQKt70R9oEMyM0EOhYnMDMLjANBgkqhkiG9w0BAQUFADA8
        MQswCQYDVQQGEwJVUzEVMBMGA1UEChMMVGhhd3RlLCBJbmMuMRYwFAYDVQQDEw1U
        aGF3dGUgU1NMIENBMB4XDTE0MDMyNTAwMDAwMFoXDTE3MDQwNTIzNTk1OVowgYox
        ...

Validation won't catch all errors, and there are plenty of ways to destroy your
YAML structure, but it's about the best one can achieve without resorting to
wanton shennanigans.

For instance, if you have a bunch of nested (concatenated) certs, Base64 encode
it and then delete a bit of it or leave some of it behind while doing a copy &
paste, you will be LUCKY if you see this...

  OpenSSL::X509::CertificateError: nested asn1 error

However, that error won't always be generated. It seems that both OpenSSL and
whatever shit Apple now uses for validating certs is extremely forgiving.

Both Puppet and `profiles` may simply accept whatever garbage you hand it and
say "Thank You!", install the profile and never tell you that it couldn't parse
the cert data.

TL;DR
As long as you keep your YAML ducks in a row and you are a careful editor, you
should now be able to embed certs in profiles as required. Boom.