prepare a validation function for raw constructors
There is a lot of dubious Puppet code in the raw constructor classes that uses the inline_template function to work around some Puppet DSL shortcomings. All of this code is being deleted and replaced with a straightforward function 'validate_raw_constructor'.

It's a weird function. It validates a Hash of Hashes (what the create_resources function expects). It does this by walking the Hash and rejecting any value that responds_to? :keys. In doing this, a successful validation results in an empty array. We do it this way so that we can return any bad data in the error message.